+++ src/main/java/com/mallang/backend/dto/PostDTO.java
@@ -1,7 +1,7 @@
-package com.mallang.backend.post.dto;
+package com.mallang.backend.dto;
 
 
-import com.mallang.backend.post.entity.PostEntity;
+import com.mallang.backend.domain.Role;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -22,7 +22,7 @@ public class PostDTO {
     private LocalDateTime postCreatedTime; //글 작성 시간
     private LocalDateTime postUpdatedTime;
 
-    public static PostDTO toPostDTO(PostEntity postEntity){
+    public static PostDTO toPostDTO(Role.PostEntity postEntity){
         PostDTO postDTO = new PostDTO();
         postDTO.setId(postEntity.getId());
         postDTO.setPostWriter(postEntity.getPostWriter());
