+++ src/main/java/com/mallang/backend/service/PostService.java
@@ -1,8 +1,8 @@
-package com.mallang.backend.post.service;
+package com.mallang.backend.service;
 
-import com.mallang.backend.post.dto.PostDTO;
-import com.mallang.backend.post.entity.PostEntity;
-import com.mallang.backend.post.repository.PostRepository;
+import com.mallang.backend.domain.Role;
+import com.mallang.backend.dto.PostDTO;
+import com.mallang.backend.repository.PostRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,15 +21,15 @@ public class PostService {
     private final PostRepository postRepository;
 
     public PostDTO save(PostDTO postDTO) {
-        PostEntity postEntity = PostEntity.toSaveEntity(postDTO);
+        Role.PostEntity postEntity = Role.PostEntity.toSaveEntity(postDTO);
         postRepository.save(postEntity);
         return postDTO;
     }
 
     public List<PostDTO> findAll() {
-        List<PostEntity> postEntityList = postRepository.findAll();
+        List<Role.PostEntity> postEntityList = postRepository.findAll();
         List<PostDTO> postDTOList = new ArrayList<>();
-        for (PostEntity postEntity : postEntityList) {
+        for (Role.PostEntity postEntity : postEntityList) {
             postDTOList.add(PostDTO.toPostDTO(postEntity));
         }
         return postDTOList;
@@ -42,9 +42,9 @@ public class PostService {
 
     @Transactional
     public PostDTO findById(Long id) {
-        Optional<PostEntity> optionalPostEntity = postRepository.findById(id);
+        Optional<Role.PostEntity> optionalPostEntity = postRepository.findById(id);
         if (optionalPostEntity.isPresent()) {
-            PostEntity postEntity = optionalPostEntity.get();
+            Role.PostEntity postEntity = optionalPostEntity.get();
             PostDTO postDTO = PostDTO.toPostDTO(postEntity);
             return postDTO;
         } else {
@@ -53,7 +53,7 @@ public class PostService {
     }
 
     public PostDTO update(PostDTO postDTO) {
-        PostEntity postEntity = PostEntity.toUpdateEntity(postDTO);
+        Role.PostEntity postEntity = Role.PostEntity.toUpdateEntity(postDTO);
         postRepository.save(postEntity);
         return findById(postDTO.getId());
     }
